Create QDplus Approximation
For Step 3) To establish an initial guess for B on n points. 

Please kindly note that the results are wrong at the moment. 

From Prof's codes for r= 0.05, q= 0.05, sigma = 0.25, K=100, tau= 0.5 the Exercise Boundary is 92.07738536421044. My result using the codes is 77.02712514058206, which is way far off at the moment.  

Let me go through them again..